+++ src/adzerk/boot_cljs.clj
@@ -72,7 +72,7 @@
             (adzerk.boot-cljs.impl/compile-cljs ~(.getPath tmp-src) ~opts))]
       (swap! core/*warnings* + (or (count warnings) 0))
       (when exception
-        (throw (util/deserialize-exception exception)))
+        (throw (util/deserialize-object exception)))
       (-> result
           (update-in [:dep-order] #(->> (conj % (:output-to opts)) (map rel-path)))
           (assoc :opts opts)))))