+++ src/adzerk/boot_cljs/impl.clj
@@ -3,6 +3,7 @@
             [boot.kahnsort :as kahn]
             [boot.pod :as pod]
             [boot.util :as butil]
+            [clojure.string :as str]
             [cljs.analyzer :as ana]
             [cljs.analyzer.api :as ana-api :refer [empty-state warning-enabled?]]
             [cljs.build.api :as build-api :refer [build inputs target-file-for-cljs-ns]]
@@ -34,7 +35,8 @@
   (->> (cljs-depdendency-graph env)
        (kahn/topo-sort)
        reverse
-       (map #(.getPath (target-file-for-cljs-ns % (:output-dir opts))))))
+       (map #(.getPath (target-file-for-cljs-ns % (:output-dir opts))))
+       vec))
 
 (defn handle-ex
   "Rethrows an interesting exception if possible or the original exception.
@@ -47,21 +49,30 @@
   (let [{:keys [type tag file line column] :as data} (util/merge-cause-ex-data e)
         ; Get the message without location info
         message (.getMessage (util/last-cause e))
-        path (util/find-original-path source-paths dirs file)
-        ; Add location info
-        message (cond
-                  column (format "ERROR: %s at file %s, line %d, column %d\n" message path line column)
-                  line (format "ERROR: %s at file %s, line %d\n" message path line)
-                  file (format "ERROR: %s at file %s\n" message path)
-                  :else message)]
+        ; Sometimes filepath is a URI
+        file (if file (str/replace file #"^file:" ""))
+        file (util/find-original-path source-paths dirs file)
+        ; Remove file info from message
+        message (if (and file (re-matches #".*in file.*" message))
+                  (first (str/split message #"\s*in file\s*"))
+                  message)
+        ; Add file info with pretty path
+        message (cond-> message
+                  line (str " at line " line)
+                  (and line column) (str ", column " column)
+                  file (str " in file " file))
+        cljs-error? (or (= :reader-exception type)
+                        (= :cljs/analysis-error tag))]
 
-    (util/serialize-exception
-      (ex-info
-        message
-        (-> data
-            (assoc :from :boot-cljs)
-            (cond-> path (assoc :file path)))
-        (.getCause e)))))
+    (ex-info
+      ;; For now, when Boot shows the message with omit-stacktrace, no newline is automatically appended
+      (if cljs-error? (str message "\n") message)
+      (-> data
+          (assoc :from :boot-cljs)
+          (cond->
+            file (assoc :file file)
+            cljs-error? (assoc :boot.util/omit-stacktrace? true)))
+      e)))
 
 (defn compile-cljs
   "Given a seq of directories containing CLJS source files and compiler options
@@ -78,13 +89,26 @@
         handler (fn [warning-type env extra]
                   (when (warning-enabled? warning-type)
                     (when-let [s (ana/error-message warning-type extra)]
-                      (let [path (util/find-original-path source-paths dirs ana/*cljs-file*)]
-                        (butil/warn "WARNING: %s %s\n" s (when (:line env)
-                                                           (str "at line " (:line env) " " path)))
-                        (swap! warnings conj {:message s
-                                              :file path
-                                              :line (:line env)
-                                              :type warning-type})))))]
+                      (let [path (if (= (-> env :ns :name) 'cljs.core)
+                                   "cljs/core.cljs"
+                                   (util/find-original-path source-paths dirs ana/*cljs-file*))
+                            warning-data {:line (:line env)
+                                          :column (:column env)
+                                          :ns (-> env :ns :name)
+                                          :file path
+                                          :type warning-type
+                                          :message s
+                                          ;; :fn-deprecated warning contains the whole
+                                          ;; analyze result as extra data, and :env value
+                                          ;; of that is unprintable.
+                                          :extra (cond-> extra
+                                                   (:fexpr extra) (update :fexpr dissoc :env))}]
+                        (butil/warn "WARNING: %s %s\n" s (if (:line env)
+                                                           (str "at line " (:line env) " " path)
+                                                           (when path
+                                                             (str "in file " path))))
+                        (butil/dbug* "%s\n" (butil/pp-str warning-data))
+                        (swap! warnings conj warning-data)))))]
     (try
       (build
         (inputs input-path)
@@ -93,7 +117,7 @@
       {:warnings  @warnings
        :dep-order (dep-order stored-env opts)}
       (catch Exception e
-        {:exception (handle-ex e source-paths dirs)}))))
+        {:exception (util/serialize-object (handle-ex e source-paths dirs))}))))
 
 (def tracker (atom nil))