+++ src/adzerk/boot_cljs/util.clj
@@ -1,10 +1,9 @@
 (ns adzerk.boot-cljs.util
   (:require [clojure.java.io :as io]
             [clojure.string :as string]
-            [boot.file :as file]
-            [clojure.walk :as walk])
-  (:import [clojure.lang ExceptionInfo]
-           [java.io File]))
+            [boot.file :as file])
+  (:import [java.util Base64]
+           [java.io ObjectInputStream ObjectOutputStream ByteArrayOutputStream ByteArrayInputStream]))
 
 (defn path->js
   "Given a path to a CLJS namespace source file, returns the corresponding
@@ -45,50 +44,25 @@
     filepath))
 
 ;;
-;; Exception serialization
+;; Object serialization
 ;;
 
-(defn safe-data [data]
-  (walk/postwalk
-    (fn [x]
-      (cond
-        (instance? File x) (.getPath x)
-        :else x))
-    data))
-
-(defn serialize-exception
-  "Serializes given exception keeping original message, stack-trace, cause stack
-   and ex-data for ExceptionInfo.
-
-   Certain types in ex-data are converted to strings. Currently this includes
-   Files."
+(defn serialize-object
+  "Serialize given Object to String using Object Streams and encode the bytes
+  as Base64 string."
   [e]
-  {:message (.getMessage e)
-   :ex-data (safe-data (ex-data e))
-   :stack-trace (mapv #(select-keys (bean %) [:className :methodName :fileName :lineNumber])
-                      (.getStackTrace e))
-   :cause (if-let [cause (.getCause e)]
-            (serialize-exception cause))})
+  (with-open [bos (ByteArrayOutputStream.)
+              out (ObjectOutputStream. bos)]
+    (.writeObject out e)
+    (.encodeToString (Base64/getEncoder) (.toByteArray bos))))
 
-(defn ->StackTraceElement [{:keys [className methodName fileName lineNumber]}]
-  (StackTraceElement. className methodName fileName lineNumber))
-
-(defn deserialize-exception
-  "Returns mocked Exception from serialized data.
-
-   Only overrides getStackTrace, which is used at least by clojure.stacktrace/print-stack-trace
-   and Boot to print the stacktrace. Some other stuff could call directly to
-   printStackTrace."
-  [{:keys [message ex-data stack-trace cause]}]
-  (let [cause (if cause (deserialize-exception cause))
-        stack-trace (into-array StackTraceElement (map ->StackTraceElement stack-trace))]
-    (if ex-data
-      (proxy [ExceptionInfo] [message ex-data cause]
-        (getStackTrace []
-          stack-trace))
-      (proxy [Throwable] [message cause]
-        (getStackTrace []
-          stack-trace)))))
+(defn deserialize-object
+  "Deserialize given Base64 encoding string using Object Streams and return the
+  Object."
+  [ba]
+  (with-open [bis (ByteArrayInputStream. (.decode (Base64/getDecoder) ba))
+              in  (ObjectInputStream. bis)]
+    (.readObject in)))
 
 (defn merge-cause-ex-data
   "Merges ex-data from all exceptions in cause stack. First value for a key is